% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ITNcoverage_scenarios.R
\name{generate_continuousDistr_SpecificAges}
\alias{generate_continuousDistr_SpecificAges}
\title{Continuous ITN Distribution (Specific Age Groups)}
\usage{
generate_continuousDistr_SpecificAges(
  halflife_weibull = 2.1,
  shape_weibull = 2,
  reach = 0.8,
  max_usage = 0.8,
  ageGroup = "U1",
  otherageGroup = "others",
  ageGroupProp = 0.035
)
}
\arguments{
\item{halflife_weibull}{Numeric. ITN half-life (years).}

\item{shape_weibull}{Numeric. Weibull shape parameter.}

\item{reach}{Numeric. Proportion of the target population reached.}

\item{max_usage}{Numeric. Maximum achievable ITN usage.}

\item{ageGroup}{Character. Name of the targeted age group.}

\item{otherageGroup}{Character. Name of the non-targeted age group.}

\item{ageGroupProp}{Numeric. Proportion of total population in target group.}
}
\value{
A data.frame containing time series of ITN usage by age group.
}
\description{
Simulates continuous annual ITN distribution targeting a specific age group,
including spillover to non-targeted groups.
}
