% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ITNcoverage_scenarios.R
\name{generate_continuousDistr_uniformAllAges}
\alias{generate_continuousDistr_uniformAllAges}
\title{Continuous ITN Distribution (Uniform Across All Ages)}
\usage{
generate_continuousDistr_uniformAllAges(
  halflife_weibull = 2.1,
  shape_weibull = 2,
  reach = 0.8,
  max_usage = 0.8,
  ageGroupProp = 0.035
)
}
\arguments{
\item{halflife_weibull}{Numeric. ITN half-life (years).}

\item{shape_weibull}{Numeric. Weibull shape parameter.}

\item{reach}{Numeric. Proportion of the population reached.}

\item{max_usage}{Numeric. Maximum achievable ITN usage.}

\item{ageGroupProp}{Numeric. Proportion of population in the targeted age group.}
}
\value{
A data.frame containing time series of total ITN usage.
}
\description{
Simulates continuous annual ITN distribution applied uniformly across
the population.
}
