% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ITNcoverage_scenarios.R
\name{generate_massCampaign_SpecificAges}
\alias{generate_massCampaign_SpecificAges}
\title{Mass Campaign ITN Distribution (Specific Age Groups)}
\usage{
generate_massCampaign_SpecificAges(
  halflife_weibull = 2.1,
  shape_weibull = 2,
  reach = 0.8,
  max_usage = 0.8,
  frequency = 3,
  ageGroup = "U5",
  otherageGroup = "others",
  ageGroupProp = 0.16,
  full_output = FALSE
)
}
\arguments{
\item{halflife_weibull}{Numeric. ITN half-life (years).}

\item{shape_weibull}{Numeric. Weibull shape parameter.}

\item{reach}{Numeric. Proportion of the target population reached.}

\item{max_usage}{Numeric. Maximum achievable ITN usage.}

\item{frequency}{Numeric. Distribution frequency (years).}

\item{ageGroup}{Character. Name of the targeted age group.}

\item{otherageGroup}{Character. Name of the non-targeted age group.}

\item{ageGroupProp}{Numeric. Proportion of total population in target group.}

\item{full_output}{Boolean. Whether all intermediate coverage variables are returned (default =FALSE)}
}
\value{
A data.frame with time series of ITN usage for target and
non-target age groups.
}
\description{
Simulates periodic mass ITN campaigns targeting a specific age group,
including spillover effects to other age groups.
}
