% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{getAverageCov_MassCampaign}
\alias{getAverageCov_MassCampaign}
\title{Average Coverage from Mass Campaign}
\usage{
getAverageCov_MassCampaign(halflife, reach, max_usage, shape_weibull = 2)
}
\arguments{
\item{halflife}{Numeric. Initial half-life of the intervention.}

\item{reach}{Numeric. Proportion of population reached.}

\item{max_usage}{Numeric. Maximum achievable usage.}

\item{shape_weibull}{Numeric. Shape parameter of Weibull decay (default = 2).}
}
\value{
Numeric. Average usage coverage.
}
\description{
Computes average coverage between years 12 and 15 for a mass campaign.
}
