% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{getCov_ContinuousSchool}
\alias{getCov_ContinuousSchool}
\title{Continuous Distribution Coverage (Age-Specific: Schoolchildren)}
\usage{
getCov_ContinuousSchool(
  halflife,
  reach,
  max_usage,
  pop6to14 = 0.27,
  prop_school_classes = 0.5,
  shape_weibull = 2
)
}
\arguments{
\item{halflife}{Numeric. Initial half-life of the intervention.}

\item{reach}{Numeric. Proportion of population reached.}

\item{max_usage}{Numeric. Maximum achievable usage.}

\item{pop6to14}{Numeric. Proportion aged 6–14 (default = 0.27).}

\item{prop_school_classes}{Numeric. Proportion attending school (default = 0.5).}

\item{shape_weibull}{Numeric. Shape parameter of Weibull decay (default = 2).}
}
\value{
List with coverage for schoolchildren and others.
}
\description{
Computes coverage for schoolchildren and others in a continuous distribution.
}
