% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{getCov_ContinuousU1}
\alias{getCov_ContinuousU1}
\title{Continuous Distribution Coverage (Age-Specific: Under-1)}
\usage{
getCov_ContinuousU1(
  halflife,
  reach,
  max_usage,
  popU1 = 0.035,
  shape_weibull = 2
)
}
\arguments{
\item{halflife}{Numeric. Initial half-life of the intervention.}

\item{reach}{Numeric. Proportion of population reached.}

\item{max_usage}{Numeric. Maximum achievable usage.}

\item{popU1}{Numeric. Proportion under 1 year (default = 0.035).}

\item{shape_weibull}{Numeric. Shape parameter of Weibull decay (default = 2).}
}
\value{
List with coverage for under-1 and others.
}
\description{
Computes coverage separately for under-1 and other age groups.
}
