% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{getCov_MassCampaignU5unif}
\alias{getCov_MassCampaignU5unif}
\title{Coverage from Mass Campaign (Uniform, Under-5 Proportion)}
\usage{
getCov_MassCampaignU5unif(
  halflife,
  reach,
  max_usage,
  popU5 = 0.16,
  shape_weibull = 2
)
}
\arguments{
\item{halflife}{Numeric. Initial half-life of the intervention.}

\item{reach}{Numeric. Proportion of population reached.}

\item{max_usage}{Numeric. Maximum achievable usage.}

\item{popU5}{Numeric. Proportion of population under 5 (default = 0.16).}

\item{shape_weibull}{Numeric. Shape parameter of Weibull decay (default = 2).}
}
\value{
Named numeric vector with updated half-life, kappa, and usage.
}
\description{
Same as \code{getCov_MassCampaign} but allows specifying the proportion of
under-5 population in a uniform-age simulation.
}
